Renamed libLERC to LERCV1
reduced number of files
