+++ gdal/frmts/mrf/LERCV1/LICENSE.TXT
