+++ gdal/frmts/mrf/LERCV1/NOTICE.TXT
