+++ gdal/frmts/mrf/LERCV1/makefile.vc
@@ -1,0 +1,22 @@
+GDAL_ROOT	=	..\..\..
+!INCLUDE $(GDAL_ROOT)\nmake.opt
+
+OBJ	= CntZImage.obj
+
+HEADERS = CntZImage.h
+
+LERCV1 = LERCV1.obj
+
+GDAL_ROOT	=	..\..\..
+
+EXTRAFLAGS      = -DLERC
+
+default:    $(LERCV1)
+
+$(OBJ):	$(HEADERS)
+
+$(LERCV1):	$(OBJ)
+        lib /out:..\$(LERCV1) $(OBJ)
+
+clean:
+	-del ..\$(LERCV1) $(OBJ)
