--- gdal/frmts/mrf/libLERC/makefile.vc
@@ -1,23 +1,0 @@
-GDAL_ROOT	=	..\..\..
-!INCLUDE $(GDAL_ROOT)\nmake.opt
-
-
-OBJ	= BitMaskV1.obj BitStufferV1.obj CntZImage.obj
-
-HEADERS = BitMaskV1.h BitStufferV1.h CntZImage.h DefinesV1.h
-
-LIBLERC = libLERC.obj
-
-GDAL_ROOT	=	..\..\..
-
-EXTRAFLAGS      = -DLERC
-
-default:    $(LIBLERC)
-
-$(OBJ):	$(HEADERS)
-
-$(LIBLERC):	$(OBJ)
-        lib /out:..\$(LIBLERC) $(OBJ)
-
-clean:
-	-del ..\$(LIBLERC) $(OBJ)
